jsonapi: refactor test helpers into base class.
